revised algorithm
