+++ README.md
@@ -1,4 +1,4 @@
-# K-Alpha Calculator: A User-Friendly Tool for Computing Krippendorf’s Alpha
+# K-Alpha Calculator —Krippendorff's Alpha Calculator: A User-Friendly Tool for Computing Krippendorf's Alpha Inter-Rater Reliability Coefficient
 
 ## Table of Contents
 1. [Introduction](#Introduction)
@@ -10,42 +10,42 @@
 
 ## Introduction
 
-Inter-coder reliability serves as a cornerstone of rigorous empirical research, particularly in the social sciences. The K-Alpha Calculator, *Krippendorff's Alpha Calculator*,  is designed to simplify the computation of Krippendorff’s Alpha, one of the most widely employed statistical measures for inter-coder reliability (Krippendorff, 2019).
-By providing a free tool with a user-friendly interface for calculating this coefficient, the K-Alpha Calculator aims to enhance research quality and reliability by offering an accessible and standalone tool for user.
+Inter-rater reliability is a cornerstone of rigorous empirical research in different contexts and domains. The K-Alpha Calculator, *Krippendorff's Alpha Calculator*,  is designed to simplify the computation of Krippendorff’s Alpha, one of the most widely employed statistical measures for inter-rater reliability (Krippendorff, 2019).
+By providing a free tool with a user-friendly interface for calculating this coefficient, the K-Alpha Calculator aims to enhance research quality and reliability by offering an accessible and standalone tool for users.
 For an in-depth review of the theoretical and mathematical underpinnings of Krippendorff's Alpha, please refer to [Krippendorff (2019)](https://doi.org/10.4135/9781071878781). For a detailed overview of The K-Alpha Calculator please refer to our [academic article](https://www.k-alpha.org/article).
 
 ---
 
 ## Software Description
 
-The K-Alpha Calculator is hosted at [k-alpha.org](https://www.k-alpha.org/), it is completely free, and its open-source code can be accessed here on [GitHub](https://github.com/davide-marchiori/k-alpha). The code behind the K-Alpha Calculator performs the calculation steps following the procedure on [Krippendorff (2019)](https://doi.org/10.4135/9781071878781). It also incorporates the adjustments and considerations outlined in the bootstrapping method as described by [Hayes and Krippendorff (2007)](https://doi.org/10.1080/19312450709336664).
+The K-Alpha Calculator is hosted at [k-alpha.org](https://www.k-alpha.org/). It is completely free, and its source code can be accessed here on [GitHub](https://github.com/davide-marchiori/k-alpha). The code behind the K-Alpha Calculator performs the calculation steps following the procedure on [Krippendorff (2019)](https://doi.org/10.4135/9781071878781). It also incorporates the adjustments and considerations outlined in the bootstrapping method as described by [Hayes and Krippendorff (2007)](https://doi.org/10.1080/19312450709336664).
 
 ### How to Use
 
 The use of K-Alpha Calculator follows a **three-step approach**:
 
-1. **Data Upload**: Users upload a data file in `.csv` format that adheres to specific formatting requirements. 
+1. **Data Upload**: Users upload a data file in `.csv` format that meets specific formatting requirements. 
  -   *File Type*: The file must be a .csv (Comma-Separated Values) format. **The separator must be a comma (,) or a semicolon (;). Tab or other separators are not allowed.**
 -   *Layout*: Data should be organized in a matrix format where:
     
-    -   Rows represent individual items to be coded.
+    -   Rows represent individual items to be rated.
         
-    -   Columns correspond to the coders providing these codes.
+    -   Columns correspond to the raters providing these rates.
         
-    -   **Code values must be represented numerically. Decimals are not allowed**
+    -   **Rate values must be represented numerically. Decimals are not allowed**
         
     -   `NA` represents a missing value.
         
--   *Content*: Each cell within this matrix should contain the code assigned by a coder to a specific item. Data should be [nominal, ordinal, interval, or ratio.](https://en.wikipedia.org/wiki/Level_of_measurement)
+-   *Content*: Each cell within this matrix should contain the rate a rater assigns to a specific item. Data should be [nominal, ordinal, interval, or ratio.](https://en.wikipedia.org/wiki/Level_of_measurement)
    
-2. **Data Type Specification**: After the data file is uploaded, the user specify the nature of the data (Nominal, Ordinal, Interval, Ratio).
+2. **Data Type Specification**: After the data file is uploaded, the user specifies the nature of the data (Nominal, Ordinal, Interval, Ratio).
 
-3. **Result Interpretation**: K-Alpha Calculator processes the uploaded data and outputs the Krippendorff's Alpha value, which reflects the level of agreement among coders. Confidence Intervals (CI) are also available with bootstrap.
-An acceptable level of reliability should be equal or above `0.80`, as suggested by [Krippendorff (2019)](https://doi.org/10.4135/9781071878781).
+3. **Result Interpretation**: K-Alpha Calculator processes the uploaded data and outputs Krippendorff's Alpha value, reflecting the agreement level among raters. Confidence Intervals (CI) are also available with bootstrap.
+An acceptable level of reliability should be equal to or above `0.80`, as suggested by [Krippendorff (2019)](https://doi.org/10.4135/9781071878781).
 
 ### Data Privacy
 
-The K-Alpha Calculator does not store any data and deletes all data upon the completion of the calculation instance.
+The K-Alpha Calculator is designed with the highest commitment to ethical data management and user privacy. It ensures that all data input by users remains exclusively within their own domain, as it operates entirely on the client side. This means the K-Alpha Calculator functions solely within the user's browser environment, without retaining, storing, or transmitting any data to external servers or storage systems.
 
 ---
 
@@ -53,7 +53,7 @@ The K-Alpha Calculator does not store any data and deletes all data upon the com
 
 ### Data Structure
 
-Consider a scenario with 5 items, 3 coders, and 2 categories represented by the values ‘1’  and ‘2’ (nominal data).​ The data file should comprise 5 rows (one for each item) and 3 columns (one for each coder). The cell values denote the codes assigned. An example representation is as follows:
+Consider a scenario with 5 items, 3 raters, and 2 categories represented by the values ‘1’  and ‘2’ (nominal data).​ The data file should comprise 5 rows (one for each item) and 3 columns (one for each rater). The cell values denote the rates assigned. An example representation is as follows:
 
 
     1,1,1
@@ -64,7 +64,7 @@ Consider a scenario with 5 items, 3 coders, and 2 categories represented by the 
 
 ### Missing Values
 
-When a coder has not assigned a code to an item, the corresponding cell should be marked as ‘NA’ (without quotation marks). For example, if coder 3 did not assign a code to item 1 and coder 2 did not assign a code to item 4, the data file should appear as follows:
+When a rater has not assigned a rate to an item, the corresponding cell should be marked as ‘NA’ (without quotation marks). For example, if rater 3 did not assign a rate to item 1 and rater 2 did not assign a rate to item 4, the data file should appear as follows:
 
     1,1,NA
     1,1,1
@@ -85,5 +85,4 @@ For additional resources, please visit [k-alpha.org](https://www.k-alpha.org/).
 
 ---
 
-For further inquiries or support, please contact [Davide Marchiori](mailto:mrcdvd77@gmail.com).
-
+For further inquiries or support, please contact us at [davmar@sam.sdu.dk](mailto:davmar@sam.sdu.dk).
