+++ src/components/MenuBootstrapIterations.jsx
@@ -21,7 +21,7 @@ export function MenuBootstrapIterations() {
   return (
     <Menu as="div" className="relative inline-block text-left">
       <div>
-        <Menu.Button className="inline-flex w-full justify-center gap-x-1.5 rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-md ring-1 ring-inset ring-gray-300 hover:bg-gray-50">
+        <Menu.Button className="inline-flex w-full justify-center gap-x-1.5 rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 ring-1 ring-inset ring-gray-300 hover:bg-gray-100">
           {sessionParams.bootIterations}
           <ChevronDownIcon
             className="-mr-1 h-5 w-5 text-gray-400"
