+++ src/components/index.js
@@ -4,4 +4,4 @@ export * from "@/components/DropZone";
 export * from "@/components/MenuBootstrapIterations";
 export * from "@/components/MenuBootstrapSampleSize";
 export * from "@/components/MenuCISize";
-export * from "@/components/Output";
+export * from "@/components/Result";
